# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RemindMe Pro API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: number;
  uid: string;
  displayName: string;
  email: string;
  /** @nullable */
  photoUrl?: string | null;
  createdAt: string;
}

export interface RegisterUserBody {
  uid: string;
  displayName: string;
  email: string;
  /** @nullable */
  photoUrl?: string | null;
}

export type ReminderType = (typeof ReminderType)[keyof typeof ReminderType];

export const ReminderType = {
  reminder: "reminder",
  note: "note",
  alarm: "alarm",
} as const;

export type ReminderPriority =
  (typeof ReminderPriority)[keyof typeof ReminderPriority];

export const ReminderPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Reminder {
  id: number;
  userId: string;
  title: string;
  content: string;
  type: ReminderType;
  /** @nullable */
  scheduledAt?: string | null;
  isCompleted: boolean;
  priority: ReminderPriority;
  createdAt: string;
  updatedAt: string;
}

export type CreateReminderBodyType =
  (typeof CreateReminderBodyType)[keyof typeof CreateReminderBodyType];

export const CreateReminderBodyType = {
  reminder: "reminder",
  note: "note",
  alarm: "alarm",
} as const;

export type CreateReminderBodyPriority =
  (typeof CreateReminderBodyPriority)[keyof typeof CreateReminderBodyPriority];

export const CreateReminderBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface CreateReminderBody {
  userId: string;
  title: string;
  content: string;
  type: CreateReminderBodyType;
  /** @nullable */
  scheduledAt?: string | null;
  priority?: CreateReminderBodyPriority;
}

export type UpdateReminderBodyPriority =
  (typeof UpdateReminderBodyPriority)[keyof typeof UpdateReminderBodyPriority];

export const UpdateReminderBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface UpdateReminderBody {
  title?: string;
  content?: string;
  /** @nullable */
  scheduledAt?: string | null;
  isCompleted?: boolean;
  priority?: UpdateReminderBodyPriority;
}

export type AlarmRepeatType =
  (typeof AlarmRepeatType)[keyof typeof AlarmRepeatType];

export const AlarmRepeatType = {
  none: "none",
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
} as const;

export interface Alarm {
  id: number;
  userId: string;
  title: string;
  scheduledAt: string;
  repeatType: AlarmRepeatType;
  isActive: boolean;
  soundEnabled: boolean;
  createdAt: string;
}

export type CreateAlarmBodyRepeatType =
  (typeof CreateAlarmBodyRepeatType)[keyof typeof CreateAlarmBodyRepeatType];

export const CreateAlarmBodyRepeatType = {
  none: "none",
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
} as const;

export interface CreateAlarmBody {
  userId: string;
  title: string;
  scheduledAt: string;
  repeatType: CreateAlarmBodyRepeatType;
  soundEnabled?: boolean;
}

export type UpdateAlarmBodyRepeatType =
  (typeof UpdateAlarmBodyRepeatType)[keyof typeof UpdateAlarmBodyRepeatType];

export const UpdateAlarmBodyRepeatType = {
  none: "none",
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
} as const;

export interface UpdateAlarmBody {
  title?: string;
  scheduledAt?: string;
  repeatType?: UpdateAlarmBodyRepeatType;
  isActive?: boolean;
  soundEnabled?: boolean;
}

export interface Group {
  id: number;
  name: string;
  description: string;
  inviteCode: string;
  ownerId: string;
  memberCount: number;
  createdAt: string;
}

export interface CreateGroupBody {
  name: string;
  description?: string;
  ownerId: string;
}

export interface UpdateGroupBody {
  name?: string;
  description?: string;
}

export interface JoinGroupBody {
  userId: string;
  inviteCode: string;
}

export type GroupMemberRole =
  (typeof GroupMemberRole)[keyof typeof GroupMemberRole];

export const GroupMemberRole = {
  owner: "owner",
  member: "member",
} as const;

export interface GroupMember {
  id: number;
  groupId: number;
  userId: string;
  displayName: string;
  role: GroupMemberRole;
  joinedAt: string;
}

export type GroupMessageMessageType =
  (typeof GroupMessageMessageType)[keyof typeof GroupMessageMessageType];

export const GroupMessageMessageType = {
  message: "message",
  reminder: "reminder",
  alarm: "alarm",
} as const;

export interface GroupMessage {
  id: number;
  groupId: number;
  senderId: string;
  senderName: string;
  content: string;
  messageType: GroupMessageMessageType;
  /** @nullable */
  scheduledAt?: string | null;
  createdAt: string;
}

export type SendGroupMessageBodyMessageType =
  (typeof SendGroupMessageBodyMessageType)[keyof typeof SendGroupMessageBodyMessageType];

export const SendGroupMessageBodyMessageType = {
  message: "message",
  reminder: "reminder",
  alarm: "alarm",
} as const;

export interface SendGroupMessageBody {
  senderId: string;
  senderName: string;
  content: string;
  messageType: SendGroupMessageBodyMessageType;
  /** @nullable */
  scheduledAt?: string | null;
}

export type RegisterTokenBodyPlatform =
  (typeof RegisterTokenBodyPlatform)[keyof typeof RegisterTokenBodyPlatform];

export const RegisterTokenBodyPlatform = {
  web: "web",
  android: "android",
  ios: "ios",
  windows: "windows",
} as const;

export interface RegisterTokenBody {
  userId: string;
  token: string;
  platform: RegisterTokenBodyPlatform;
}

export interface TokenResponse {
  success: boolean;
  message: string;
}

export interface SendNotificationBody {
  groupId: number;
  title: string;
  body: string;
  senderId: string;
}

export interface NotificationResult {
  sent: number;
  failed: number;
}

export interface DashboardSummary {
  totalReminders: number;
  pendingReminders: number;
  completedReminders: number;
  totalAlarms: number;
  activeAlarms: number;
  totalGroups: number;
  upcomingCount: number;
}

export type ListRemindersParams = {
  userId: string;
};

export type ListAlarmsParams = {
  userId: string;
};

export type ListGroupsParams = {
  userId: string;
};

export type GetDashboardSummaryParams = {
  userId: string;
};