# lib/api-zod/src/generated/types/createAlarmBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RemindMe Pro API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateAlarmBodyRepeatType } from "./createAlarmBodyRepeatType";

export interface CreateAlarmBody {
  userId: string;
  title: string;
  scheduledAt: string;
  repeatType: CreateAlarmBodyRepeatType;
  soundEnabled?: boolean;
}