# lib/api-zod/src/generated/types/createGroupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RemindMe Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGroupBody {
  name: string;
  description?: string;
  ownerId: string;
}